Use tree of bools instead of list of errors
This is a lot easier than having to flatten and wrap another layer of
context at every level. You can always print it differently if you
want.

One disadvantage is that the other style allows you to allocate less,
because the iterators are lazily evaluated. If you never need to store
the errors, you never need to allocate storage for them. For now, I
think this is not a concern, because memory usage isn't an issue we're
worried about. If this becomes an issue, we can either switch back, or
we can come up with some kind of lazy tree data structure.

Another disadvantage is we are in some ways duplicating the structure
of the original input. You could imagine parametrizing the input over
some kind of annotation type, where for the original version we parse
that type is (). Then we transform it to an annotated version. This is
probably the Right Way™ when we have time.

It's worth seeing how people handle ASTs and typechecking, because in
many ways this is similar and that's where I got the above idea, so if
they have other solutions, those also might be applicable.