Check 0/1 proofs using status and hash spec
Port the checking of 0/1 CP proofs for selections to use the new
architecture of producing structured statuses, and using hash specs to
inject the necessary information into proofs for them to check their
own hashes. This shows how parametrizing over the provided values in
the hash spec allows the disj::Proof to provide different values than
a regular Proof.